Adding a base class for password requirements used through the application
The PasswordRequirements has constants to represent the current password options. These are used to set the Identity options and the validation annotation on RegisterViewModel. A read-only property was also created that constructs a user friendly password guideance string from the settings. This will update should there be a need to amend the settings in the future. The RegisterViewModel was not being passed into the view by the AccountController so that was changed to support the guideance in the view razor.