Use existing postal code for location or add new one.
When adding or editing an Organization, the address must contain a
postal code that has been added before (ie. in the PostalCodeGeo table).
The changes to OrganizationEditCommandHandler fetch  the added/edited
postal code OR adds the new postal code if not found. The unit tests in
OrganizationEditCommandHandlerTests test that changes to a location's
city/state are NOT copied over to exiting PostalCodeGeo records but ARE
copied to new ones.