Renaming activities to events in all cs code.
This includes class names, properties, fields and method names. This includes renaming and updating tests.

This change also includes database model updates and a migrations to update dev databases.

IMPORTANT! This migration involves some dropped tables and is destructive.

View folders and names have also been updated to ensure the site functions based on the view location conventions. Other cs files and folders have not been changed in this commit and will form their own commit and PR later to try and limit the complexity of reviewing this PR.