Fixing error where closed tasks can be volunteered for.
Fixes issue #719

NOTE: Must merge after PR #722 and PR #723 as it is based on the renaming excercise. I may need to rebase this if/when those are accepted.

Primary I have hidden the volunteer buttons for tasks which are closed. In addition I have hardened the API and handler to fail if a closed task id is sent.

Changes made:
Added bool isClosed calculated property to the task model (not mapped)
Passed the isClosed value to the TaskViewModel
Used the isClosed property to hide the volunteer buttons for closed tasks
Updated TaskSignupHandler to check for closed tasks and issue error in result
Also added checks for Event and Task not found scenarios
Updated register method for TaskAPI to handle and output correct JSON for the various error responses
Added tests for all changes with expected behaviours