Creating EditEventModel to avoid validation failures (#854)
Prior to this commit we were using the EventDetailsModel on the create/edit pages. However this model now contains extra properties, including a new Itinerary model which also have validation attributes. These were getting checked and failing required validation since when creating/editing an event we are not populating these properties.

It is therefore cleaner to have a model responsible for the event edit properties seperated from the more complex details model.