Finish test coverage for TaskStatusChangeHandlerAsync (#855)
* #787 TaskStatusChangeHandlerAsync base behavior tests
Added unit tests for testing the basic behavior of the TaskStatusChangeHandlerAsync request handler.
- Verify Assigned status change
- Verify change from Assigned to Accepted
- Verify change from Assigned to Rejected
- Verify change from Accepted to Completed
- Verify change from Accepted to CanNotComplete

* Refactored tests

Refactored unit tests to test all TaskSignup properties that get changed
when changing the signup status. I changed the 'VolunteerAssignedTask' and
'VolunteerAcceptsTask' tests.

* Updated remaining tests

Updated remaining tests based on the comments in
PR #836

* test transition task status from 'CanNotComplete' to 'Accepted'

Added a unit test for testing the task status transition from the
'CanNotComplete' state to the 'Accepted' state as part of the Tests for Success
task list on #787

* Finished tests for success checklist

Finished the 'Tests for Success' checklist on #787

* Finished tests for failure checklist

As part of issue #787, finished the tests for failure
checklist.