Updating the fix per feedback from the EF team.
See https://github.com/aspnet/EntityFramework/issues/6217

This change adds the itinerary earlier in the flow (when we have not retrived an existing one) so that EF tracking is correct and we can then simply just call SaveChanges without the need to add or update.

An existing entitiy will get added to the change tracker and fields will be marked as updated. If no existing record matched the id then a new one is added and marked as such in the change tracker.