fixed the me controller test
refactored the account controller test to use a mock helper to generate
the objects that are needed so that the create object methods could be
reused for the me controller test.

fixed the failing test for the me controller.

added test to validate that the routes are correct for the controller and
the login method

https://github.com/HTBox/allReady/issues/1097