Organization Admin: moved edit btn to left
On the Organization admin page created an action column and moved the edit
button into it.  This made it match how the rest of the site worked.

To make the UI look right, I moved the logo to before the delete button.
Also, from a UX perspective it felt like the name was more important than
the logo was so the name should have been the 1st column instead of the
logo.