Create Skill: fixed form indent & aligned btns
To fix the form indent, I changed the labels to col-sm-1 and the input elements to col-sm-11

From a UX perspective, the buttons should be under the input elements
since that it where you eyes are at when you are done filling out the last
input element.  I added the buttons to a form-group and div with a col-sm-offset-1 just like the bootstrap docs horizontal form element.

changed the Html.BeginForm call to Html.BeginForm(null, null, FormMethod.Post, new { @class = "form-horizontal" }) instead of having a div with the form-horizontal class on it.