Fixed DateTime pickers
The date time pickers were incorrectly shifting the datetime to the
user's local timezone instead of displaying in the timezone of the
campaign/event. We could have fixed thsi by changing the input type to
"datetime-local" but that causes some conflict with the brower's
implementation of the datetime picker. Avoiding this by specifying the
input type as "text" and using the custom bootstrap datetime picker