Fixing issues when volunteering for a task: (#1724)
Fixing a null reference that occurs when volunteering for a campaign that was created with the seed data without a campaign contact. The first ?. can result in a null, thus need the second ?.
Adding a defaulty (empty) list of CampaignContacts to the Campaign EF model to prevent InvalidOperation occuring in an EF query when getting tasks.