Converting HomeController pages/action to Razor pages
NOTE: We were doing some interesting stuff to work out the current page in order to modify the menus. I had to remove this for now since it doesn't work without some re-work.

I felt it was better to get the example of Razor pages in and revisit the menu highlighting later since it's not a critical function.